+++ client/src/App.tsx
@@ -1,21 +1,17 @@
 import { Route, BrowserRouter } from "react-router-dom";
 
 import Home from "./Components/Home/Home";
-import SellerProfile from "./Components/SellerProfile/SellerProfile";
+import DashboardSeller from "./Components/DashboardSeller/DashboardSeller";
 
 function App() {
-    return (
-        <BrowserRouter>
-            <div>
-                <Route exact path="/" render={() => <Home />} />
-                <Route
-                    exact
-                    path="/sellerprofile"
-                    render={() => <SellerProfile />}
-                />
-            </div>
-        </BrowserRouter>
-    );
+  return (
+    <BrowserRouter>
+      <div>
+        <Route exact path="/" render={() => <Home />} />
+        <Route exact path="/sellerprofile" render={() => <DashboardSeller />} />
+      </div>
+    </BrowserRouter>
+  );
 }
 
 export default App;
