+++ client/src/Components/DashboardSeller/DashboardSeller.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import SellerNavbar from "./NavbarSeller";
+
+export default function DashboardSeller() {
+  return (
+    <div>
+      <SellerNavbar />
+      <div>
+        <h2>Aca se muestra la info o editor</h2>
+      </div>
+    </div>
+  );
+}
