--- client/src/Components/SellerProfile/SellerProfile.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import SellerNavbar from "./SellerNavBar";
-
-export default function SellerProfile() {
-    return (
-        <div>
-            <SellerNavbar />
-            <div>
-                <h2>Aca se muestra la info o editor</h2>
-            </div>
-        </div>
-    );
-}
